[agl][log] Improve support for logging without PmLog.
:Release Notes:
Extend log support without PmLog.

:Detailed Notes:
The replacement for PmLog did only show debug level messages. This patch
adds an implementation also for all the other levels, properly showing
the PMLOGKS and PMLOGKFV macros too.

:Testing Performed:
Logs are properly shown in journalctl for all the other levels.
A fix is required for QtLess as many lines were broken and untested
because of the suppression of the log lines.

:QA Notes:

:Issues Addressed: